+++ NEWS.md
@@ -10,6 +10,7 @@ New features(CLI, Configs):
 + Add `PHAN_DISABLE_PROGRESS_BAR` environment variable to disable progress bar by default unless the CLI arg `--progress-bar` is used.
 
 New features(Analysis):
++ Make inferred real types more accurate for equality/identity/instanceof checks.
 + Make issue suggestions (and CLI suggestions) for completions of prefixes case-insensitive.
 + Support `@seal-properties` and `@seal-methods` as aliases of `@phan-forbid-undeclared-magic-properties` and `@phan-forbid-undeclared-magic-methods`
 + More aggressively infer real types of array destructuring(e.g. `[$x] = expr`) and accesses of array dimensions (e.g. `$x = expr[dim]`) (#3481)